Clip calculated SPI clock divisor to the allowable range.
The SPI clock divisor is an eight bit field. The old method would simply
take the lower eight bits as the divisor.  This meant that if you asked
for a lower clock speed than possible, it would wrap around to a high
speed.

For example, for a 96MHz peripheral clock a desired speed of 300kHz
would need a divisor of 320. Taking the lower eight bits results in a
divisor of 64, and a SPI clock of 1.5MHz. Clipping the divisor to the
maximum 255 results in the lowest achievable clock speed (~377kHz in
this example).